+++ src/simmate/database/base_data_types/base.py
@@ -19,8 +19,9 @@ from pathlib import Path
 
 import pandas
 import yaml
-from django.db import models  # , transaction
+from django.db import models  # see comment below
 from django.db import models as table_column
+from django.urls import reverse
 from django.utils.module_loading import import_string
 from django.utils.timezone import datetime
 from django_filters import rest_framework as django_api_filters
@@ -1368,6 +1369,10 @@ class DatabaseTable(models.Model):
     # Methods that link to the website UI
     # -------------------------------------------------------------------------
 
+    html_template_table: str = None
+    html_template_entry: str = None
+    # experimental override for templates using by the Data Explorer app
+
     @classmethod
     @property
     def url_table(self) -> str:
@@ -1383,3 +1388,15 @@ class DatabaseTable(models.Model):
         Provides the URL link to the database entry page in the Simmate website
         """
         return f"{self.url_table}/{self.id}"
+
+    def get_absolute_url(self):
+        """
+        Provides the full URL link to the database entry page in the Simmate website.
+
+        This is for use in template views, whereas user may find the `url`
+        property easier to work with
+        """
+        return reverse(
+            "data_explorer:entry-detail",
+            kwargs={"provider_name": self.table_name, "pk": self.pk},
+        )