+++ src/simmate/engine/execution/executor.py
@@ -4,8 +4,10 @@ import logging
 from datetime import timedelta
 
 import cloudpickle  # needed to serialize Prefect workflow runs and tasks
+import pandas
 from django.utils import timezone
 from rich import print
+from rich.progress import track
 
 from simmate.configuration.django.settings import DATABASE_BACKEND
 from simmate.engine.execution.database import WorkItem
@@ -122,20 +124,23 @@ class SimmateExecutor:
 
     @classmethod
     def delete(cls, tags: list[str], confirm: bool = False):
-        # no tags means delete all
-        if not tags:
-            cls.delete_all(confirm=confirm)
-
         # Make sure the user ment to do this, otherwise raise an exception
         if not confirm:
             raise Exception(
-                "Are you sure you want to do this? This deletes all of your queue "
+                "Are you sure you want to do this? This deletes your queue "
                 "data and you can't get it back. If so, run this method again "
                 "with confirmation."
             )
 
-        # Filtering jobs with specific tags and delete all jobs that you pulled
-        jobs = WorkItem.objects.filter_by_tags(tags=tags).all()
+        if tags:
+            # Filtering jobs with specific tags and delete all jobs that you pulled
+            jobs = WorkItem.objects.filter_by_tags(tags=tags).all()
+
+        # no tags means delete all without ANY tags. If the user meant to delete
+        # all entries, then they should be using delete_all instead
+        else:
+            jobs = WorkItem.objects.filter(tags=[]).all()
+
         jobs.delete()
 
     @staticmethod
@@ -187,28 +192,150 @@ class SimmateExecutor:
                 print(f"{job.id} | {error}")
 
     @staticmethod
-    def show_stats() -> int:
-        npending = WorkItem.objects.filter(status="P").count()
-        nrunning = WorkItem.objects.filter(status="R").count()
-        ncanceled = WorkItem.objects.filter(status="C").count()
-        nfinished = WorkItem.objects.filter(status="F").count()
-        nerrored = WorkItem.objects.filter(status="E").count()
+    def get_stats(tags: list[str] = []) -> dict:
+        query = WorkItem.objects.all()
+        if tags:
+            query = query.filter_by_tags(tags=tags)
+
+        npending = query.filter(status="P").count()
+        nrunning = query.filter(status="R").count()
+        ncanceled = query.filter(status="C").count()
+        nfinished = query.filter(status="F").count()
+        nerrored = query.filter(status="E").count()
 
         if nfinished:
             error_percent = (nerrored / (nerrored + nfinished)) * 100
         else:
             error_percent = 0
 
-        nrunning_long = WorkItem.objects.filter(
+        nrunning_long = query.filter(
             status="R",
             updated_at__lte=timezone.now() - timedelta(days=1),
         ).count()
 
-        print(f"PENDING:   {npending}")
-        print(f"RUNNING:   {nrunning} ({nrunning_long} for +24hrs)")
-        print(f"FINISHED:  {nfinished}")
-        print(f"ERRORED:   {nerrored} ({error_percent:.2f}%)")
-        print(f"CANCELED:  {ncanceled}")
+        return {
+            "npending": npending,
+            "nrunning": nrunning,
+            "ncanceled": ncanceled,
+            "nfinished": nfinished,
+            "nerrored": nerrored,
+            "error_percent": error_percent,
+            "nrunning_long": nrunning_long,
+        }
+
+    @classmethod
+    def show_stats(cls, tags: list[str] = []):
+        stats = cls.get_stats(tags=tags)
+        print(f"PENDING:   {stats['npending']}")
+        print(f"RUNNING:   {stats['nrunning']} ({stats['npending']} for +24hrs)")
+        print(f"FINISHED:  {stats['nfinished']}")
+        print(f"ERRORED:   {stats['nerrored']} ({stats['error_percent']:.2f}%)")
+        print(f"CANCELED:  {stats['ncanceled']}")
+
+    @classmethod
+    def show_stats_detail(
+        cls,
+        tags: list[str] = [],
+        recent: float = None,
+    ):
+        query = WorkItem.objects.all()
+
+        if tags:
+            query = query.filter_by_tags(tags=tags)
+        if recent:
+            query = query.filter(
+                updated_at__gte=timezone.now() - timedelta(hours=recent),
+            )
+
+        logging.info("Loading data...")
+        all_items = query.values(
+            "status",
+            "tags",
+        ).to_dataframe()
+
+        all_items["tags"] = all_items.tags.apply(tuple)
+        unique_tags = set([t for g in all_items.tags.unique() for t in g])
+        logging.info(f"Found {len(unique_tags)} unique tags")
+
+        # sort by type of tag then aphlabetical
+        unique_tags = list(unique_tags)
+        unique_tags.sort(key=lambda item: item.count("."))
+
+        logging.info("Breaking down stats for each tag...")
+        tag_data = []
+        for tag in track(unique_tags):
+            stats = cls.get_stats(tags=[tag])
+            tag_data.append(
+                [
+                    tag,
+                    stats["npending"],
+                    stats["nrunning"],
+                    stats["nrunning_long"],
+                    stats["nfinished"],
+                    stats["ncanceled"],
+                    stats["nerrored"],
+                    stats["error_percent"],
+                ]
+            )
+
+        # Add the totals count last
+        stats = cls.get_stats()
+        tag_data.append(
+            [
+                "(--ALL WORKITEMS--)",
+                stats["npending"],
+                stats["nrunning"],
+                stats["nrunning_long"],
+                stats["nfinished"],
+                stats["ncanceled"],
+                stats["nerrored"],
+                stats["error_percent"],
+            ]
+        )
+
+        tag_data = pandas.DataFrame(
+            tag_data,
+            columns=[
+                "tag",
+                "pending (#)",
+                "running (#)",
+                "running >24hrs (#)",
+                "finished (#)",
+                "canceled (#)",
+                "errored (#)",
+                "errored (%)",
+            ],
+        )
+        print(tag_data.to_markdown(index=False))
+        print("\n**keep in mind that workitems can have multiple tags")
+
+    @staticmethod
+    def show_workitems(
+        tags: list[str] = [],
+        status: str = None,
+        recent: float = None,
+    ):
+        # recent --> "only entries updated within N hours ago"
+
+        query = WorkItem.objects.order_by("created_at")
+
+        if tags:
+            query = query.filter_by_tags(tags=tags)
+        if status:
+            query = query.filter(status=status)
+        if recent:
+            query = query.filter(
+                updated_at__gte=timezone.now() - timedelta(hours=recent),
+            )
+
+        all_items = query.values(
+            "id",
+            "created_at",
+            "updated_at",
+            "status",
+            "tags",
+        ).to_dataframe()
+        print(all_items.to_markdown(index=False))
 
     # -------------------------------------------------------------------------
     # Extra methods to add if I want to be consistent with other Executor classes